+++ src/_h5ai/private/php/core/class-util.php
@@ -105,8 +105,7 @@ class Util {
     }
 
     public static function filesize(Context $context, string $path): ?int {
-        $withFoldersize = $context->query_option('foldersize.enabled', false);
-        $withDu = $context->get_setup()->get('HAS_CMD_DU') && $context->query_option('foldersize.type', null) === 'shell-du';
+        [$withFoldersize, $withDu] = $context->foldersize_mode();
         return Filesize::getCachedSize($path, $withFoldersize, $withDu);
     }